various updates
switching to id rather than kafkaClusterId
using jboss logging
using mutiny for async rest
consolidating ManagedKafkaSync logic for async
using string lookup keys

+++ agent-sync/src/main/java/org/bf2/sync/ManagedKafkaSync.java
@@ -16,8 +16,7 @@ import org.bf2.operator.resources.v1alpha1.ManagedKafkaStatusBuilder;
 import org.bf2.sync.controlplane.ControlPlane;
 import org.bf2.sync.informer.LocalLookup;
 import org.eclipse.microprofile.context.ManagedExecutor;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
+import org.jboss.logging.Logger;
 
 import io.fabric8.kubernetes.client.KubernetesClient;
 import io.fabric8.kubernetes.client.dsl.MixedOperation;
@@ -38,7 +37,8 @@ import io.quarkus.scheduler.Scheduled;
 @ApplicationScoped
 public class ManagedKafkaSync {
 
-    private static final Logger log = LoggerFactory.getLogger(ManagedKafkaSync.class);
+    @Inject
+    Logger log;
 
     @Inject
     KubernetesClient client;
@@ -62,20 +62,23 @@ public class ManagedKafkaSync {
         managedKafkaResources = client.customResources(crdContext, ManagedKafka.class, ManagedKafkaList.class);
     }
 
+    /**
+     * Update the local state based upon the remote ManagedKafkas
+     * The strategy here is to take a pass over the list and find any deferred work
+     * Then execute that deferred work using the {@link ManagedExecutor} but with
+     * a refresh of the state to ensure we're still acting appropriately.
+     */
     public void syncKafkaClusters(List<ManagedKafka> remoteManagedKafkas, Executor executor) {
         Map<String, ManagedKafka> remotes = new HashMap<>();
 
         for (ManagedKafka remoteManagedKafka : remoteManagedKafkas) {
-            remotes.put(remoteManagedKafka.getKafkaClusterId(), remoteManagedKafka);
+            remotes.put(remoteManagedKafka.getId(), remoteManagedKafka);
             ManagedKafkaSpec remoteSpec = remoteManagedKafka.getSpec();
             assert remoteSpec != null;
 
-            // TODO: account for namespaces.
-            // eventually the remote metadata should be ignored - and we should assign /
-            // create a local namespace as needed
-            String namespace = remoteManagedKafka.getMetadata().getNamespace();
-
-            ManagedKafka existing = lookup.getLocalManagedKafka(remoteManagedKafka);
+            // TODO this requires the namespace is already set - this will change
+            // see also the create method
+            ManagedKafka existing = lookup.getLocalManagedKafka(Cache.metaNamespaceKeyFunc(remoteManagedKafka));
 
             // take action based upon differences
             // this is really just seeing if an instance needs created and the delete flag
@@ -86,10 +89,7 @@ public class ManagedKafkaSync {
                     controlPlane.addManagedKafka(remoteManagedKafka);
 
                     executor.execute(() -> {
-                        log.debug("Creating ManagedKafka {}", Cache.metaNamespaceKeyFunc(remoteManagedKafka));
-                        // TODO: create namespace when needed
-
-                        managedKafkaResources.create(remoteManagedKafka);
+                        reconcile(remoteManagedKafka.getId(), Cache.metaNamespaceKeyFunc(remoteManagedKafka));
                     });
                 } else {
                     // we've successfully removed locally, but control plane is not aware
@@ -98,27 +98,23 @@ public class ManagedKafkaSync {
                     // doesn't need to be async as the control plane call is async
                     ManagedKafkaStatusBuilder statusBuilder = new ManagedKafkaStatusBuilder();
                     statusBuilder.addNewCondition().withType("InstanceDeletionComplete").endCondition();
-                    controlPlane.updateKafkaClusterStatus(statusBuilder.build(), remoteManagedKafka.getKafkaClusterId());
+
+                    // fire and forget - if it fails, we'll retry on the next poll
+                    controlPlane.updateKafkaClusterStatus(statusBuilder.build(), remoteManagedKafka.getId());
                 }
             } else if (remoteSpec.isDeleted() && !existing.getSpec().isDeleted()) {
-                controlPlane.removeManagedKafka(existing);
-                // mark the local as deleted
                 executor.execute(() -> {
-                    managedKafkaResources.inNamespace(existing.getMetadata().getNamespace())
-                            .withName(existing.getMetadata().getName()).edit(mk -> {
-                                mk.getSpec().setDeleted(true);
-                                return mk;
-                            });
-                    // the agent will handle the deletion from here
+                    reconcile(remoteManagedKafka.getId(), Cache.metaNamespaceKeyFunc(existing));
                 });
             } else if (!remoteSpec.isDeleted() && existing.getSpec().isDeleted()) {
-                // TODO: seems like a problem / resurrection
+                // TODO: seems like a problem / resurrection - should not happen
+                log.warnf("Ignoring ManagedKafka %s %s that wants to come back to life", existing.getId(), Cache.metaNamespaceKeyFunc(existing));
             }
         }
 
         // process final removals
         for (ManagedKafka local : lookup.getLocalManagedKafkas()) {
-            if (remotes.get(local.getKafkaClusterId()) != null) {
+            if (remotes.get(local.getId()) != null) {
                 continue;
             }
 
@@ -128,10 +124,65 @@ public class ManagedKafkaSync {
                 // TODO: seems bad
             }
 
-            log.debug("Deleting ManagedKafka {}", Cache.metaNamespaceKeyFunc(local));
+            executor.execute(() -> {
+                reconcile(null, Cache.metaNamespaceKeyFunc(local));
+            });
+        }
+
+    }
+
+    /**
+     * Final sync processing of the remote vs. local state
+     */
+    void reconcile(String remoteId, String localMetaNamespaceKey) {
+        ManagedKafka local = null;
+        if (localMetaNamespaceKey != null) {
+            //refresh the local
+            local = lookup.getLocalManagedKafka(localMetaNamespaceKey);
+        }
+
+        ManagedKafka remote = null;
+        if (remoteId != null) {
+            //refresh the remote
+            remote = controlPlane.getManagedKafka(remoteId);
+        }
+
+        if (local == null && remote == null) {
+            return; //nothing to do
+        }
+
+        if (local == null) {
+            if (!remote.getSpec().isDeleted()) {
+                create(remote);
+            }
+        } else if (remote == null) {
+            log.debugf("Deleting ManagedKafka %s %s", local.getId(), Cache.metaNamespaceKeyFunc(local));
+
             managedKafkaResources.inNamespace(local.getMetadata().getNamespace())
                     .withName(local.getMetadata().getName()).delete();
+        } else if (remote.getSpec().isDeleted() && !local.getSpec().isDeleted()) {
+            log.debugf("Initiating Delete of ManagedKafka %s %s", remote.getId(), Cache.metaNamespaceKeyFunc(remote));
+            // mark the local as deleted
+            managedKafkaResources.inNamespace(local.getMetadata().getNamespace())
+                    .withName(local.getMetadata().getName()).edit(mk -> {
+                        mk.getSpec().setDeleted(true);
+                        return mk;
+                    });
+            controlPlane.removeManagedKafka(local);
+            // the operator will handle the deletion from here
         }
+    }
+
+    void create(ManagedKafka remote) {
+        // TODO: account for namespaces.
+        // eventually the remote metadata should be ignored (or unset) - and we should assign /
+        // create a local namespace as needed
+        // for now we're assuming remote.getMetadata().getNamespace() is set;
+
+        log.debugf("Creating ManagedKafka %s %s", remote.getId(), Cache.metaNamespaceKeyFunc(remote));
+
+        // TODO: there may be additional cleansing, setting of defaults, etc. on the remote instance before creation
+        managedKafkaResources.create(remote);
     }
 
     @Scheduled(every = "{poll.interval}")
